DataSys(cloudver)/requirements.txt
streamlit
pandas
numpy
plotly
altair
google-cloud-bigquery
google-cloud-storage
google-auth
db-dtypes
pydantic
watchdog
streamlit-aggrid
streamlit-authenticator
pyyaml
google-cloud-bigquery-storage
